C-api: refine interface on Selectable classes (#945)
Previously, methods like hasData/hasCachedData were exposed through
the FFI, but they are only useful for swss::Select which is not
exposed, so those were deleted. Methods *_getFd were added so that
ffi code may select on those fds separately from swss::Select (we
need this in Rust for tokio::AsyncFd). Comments were added to better
explain the intended use - Either call readData with a timeout to
block in place, or use getFd to select on the fd and then call readData
with a zero timeout to reset the fd.

Co-authored-by: erer1243 <erer1243@users.noreply.github.com>