C API: improve performance by not memcpy-ing string DB values across ffi boundary  (#935)
This introduces a couple of new FFI types, representing owned C++ strings and arrays. We use those owned strings for database values, so they do not need to be memcpy'd across ffi boundaries. This is in response to Riff's concerns that database values which hamgrd will be reading may be large and expensive to memcpy.

Partner pr to sonic-net/sonic-dash-ha#11

Co-authored-by: erer1243 <erer1243@users.noreply.github.com>